World fix
i fixed the error that shown itself when starting the world scene. animations only wokr for the dummy class. i will continue to finish the animations for the characters. Some more head name are included for the customization. I'm working on the headless sprites, and we need better head alignment for the animations aswell. The PUI is set for the start of attacks and skills. as well as statuses. once we get hitboxes attacks and stuff we can start working on the multiplayer networking. for testing the attacks use a dummy enemy. Also some trash files that we don't need were thrown out.